All repositories refactoring for MyBatis

+++ tm-server/src/main/java/ru/stepanenko/tm/service/ProjectService.java
@@ -1,35 +1,39 @@
 package ru.stepanenko.tm.service;
 
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 import ru.stepanenko.tm.api.service.IProjectService;
 import ru.stepanenko.tm.api.repository.IProjectRepository;
 import ru.stepanenko.tm.entity.Project;
 import ru.stepanenko.tm.enumerate.Status;
-import ru.stepanenko.tm.util.ComparatorUtil;
+import ru.stepanenko.tm.util.ParameterValidator;
 import ru.stepanenko.tm.util.EnumUtil;
 import ru.stepanenko.tm.util.StringValidator;
 
 import java.util.Collection;
 import java.util.Date;
 
-public final class ProjectService extends AbstractEntityService<Project, IProjectRepository> implements IProjectService {
+public final class ProjectService implements IProjectService {
 
+    @NotNull final IProjectRepository repository;
     public ProjectService(@NotNull final IProjectRepository projectRepository) {
-        super(projectRepository);
+        this.repository = projectRepository;
     }
 
     @Override
     public Project create(@NotNull final String name, @NotNull final String description, @NotNull final String userId) {
         if (!StringValidator.validate(name, description, userId)) return null;
-        return repository.persist(new Project(name, description, userId));
+        @NotNull Project project = new Project(name, description, userId);
+        repository.persist(project);
+        return findOne(project.getId());
     }
 
     @Override
     public Project edit(@NotNull String id, @NotNull String name, @NotNull String description, @NotNull String status, @NotNull String userId) {
         if (!StringValidator.validate(id, name, description, status)) return null;
         if (EnumUtil.stringToStatus(status) == null) return null;
-        @NotNull
-        Project project = findOne(id, userId);
+        @Nullable final Project project = findOne(id, userId);
+        if (project==null) return null;
         project.setName(name);
         project.setDescription(description);
         project.setStatus(EnumUtil.stringToStatus(status));
@@ -38,7 +42,8 @@ public final class ProjectService extends AbstractEntityService<Project, IProjec
         } else {
             project.setDateEnd(null);
         }
-        return repository.merge(project);
+        repository.merge(project);
+        return findOne(project.getId());
     }
 
     @Override
@@ -50,7 +55,40 @@ public final class ProjectService extends AbstractEntityService<Project, IProjec
     @Override
     public Project remove(@NotNull String id, @NotNull String userId) {
         if (!StringValidator.validate(id, userId)) return null;
-        return repository.remove(id, userId);
+        @Nullable final Project project = findOne(id, userId);
+        if(project==null) return null;
+        repository.remove(id, userId);
+        return project;
+    }
+
+    @Override
+    public void clear() {
+        repository.removeAll();
+    }
+
+    @Override
+    public Project findOne(@NotNull String id) {
+        if (!StringValidator.validate(id)) return null;
+        return repository.findOne(id);
+    }
+
+    @Override
+    public Project remove(@NotNull String id) {
+        if (!StringValidator.validate(id)) return null;
+        @Nullable final Project project = findOne(id);
+        if(project==null) return null;
+        repository.remove(id);
+        return project;
+    }
+
+    @Override
+    public void recovery(@NotNull Collection<Project> collection) {
+        repository.recovery(collection);
+    }
+
+    @Override
+    public Collection<Project> findAll() {
+        return repository.findAll();
     }
 
     @Override
@@ -66,11 +104,11 @@ public final class ProjectService extends AbstractEntityService<Project, IProjec
     }
 
     @Override
-    public Collection<Project> sortAllByUserId(@NotNull String id, @NotNull String comparator) {
-        if (!StringValidator.validate(id, comparator)) return null;
-        if ("order".equals(comparator)) return findAllByUserId(id);
-        if (ComparatorUtil.getProjectComparator(comparator) == null) return null;
-        return repository.sortAllByUserId(id, ComparatorUtil.getProjectComparator(comparator));
+    public Collection<Project> sortAllByUserId(@NotNull String id, @NotNull String parameter) {
+        if (!StringValidator.validate(id, parameter)) return null;
+        if (!ParameterValidator.validate(parameter)) return null;
+        if ("order".equals(parameter)) return findAllByUserId(id);
+        return repository.sortAllByUserId(id, parameter);
     }
 
     @Override
